validacion de eliminacion cilindros

+++ sections/cilindros/index.php
@@ -4,14 +4,23 @@ include("../../db.php");
 if (isset($_GET['txtID'])) {
     $txtID = (isset($_GET['txtID'])) ? $_GET['txtID'] : "";
 
-    $sentencia = $conexion->prepare("DELETE FROM cilindros WHERE id=:id");
+    $sentencia = $conexion->prepare("SELECT * FROM puntoventa_bodega WHERE id_cilindros=:id");
     $sentencia->bindParam(":id", $txtID);
-
     $sentencia->execute();
+    $registro_recuperado = $sentencia->fetchAll(PDO::FETCH_ASSOC);
 
-    $mensaje="Registro eliminado";
-    $icono="success";
-    header("Location:index.php?mensaje=".$mensaje."&icono=".$icono);
+    if(!empty($registro_recuperado)){
+        $mensaje="Imposible eliminar: cilindro tiene relaciones con puntos de ventas y bodegas";
+        $icono="error";
+        header("Location:index.php?mensaje=".$mensaje."&icono=".$icono);
+    }else{
+        $sentencia = $conexion->prepare("DELETE FROM cilindros WHERE id=:id");
+        $sentencia->bindParam(":id", $txtID);
+        $sentencia->execute();
+        $mensaje="Registro eliminado";
+        $icono="success";
+        header("Location:index.php?mensaje=".$mensaje."&icono=".$icono);
+    }
 }
 
 $sql = $conexion->prepare("SELECT * FROM cilindros;");
